+++ docs/changelog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this section.
 ## [1.0.0-rc.4] - Unreleased
 
 - fix(process): Unref forced termination timer by passing `persistant: false` to delay, to prevent it for keeping the main process alive for the full duration of the timeout
+- feature(web-interface): Add history to web interface log view
 
 ## [1.0.0-rc.3] - 2023-07-10